use fullscreen window for every window replace sinlge fullscreen to exachange because in some PC, the exachange between two swapchain presents not promised in same frame so the blinking can be happened will, fullscreen window for every window will consume many GPU memory. But I want the smooth resizing and moving!!!

+++ assets/shader.hlsl
@@ -20,7 +20,7 @@ struct PSParameter
 struct Constants
 {
   uint2    window_extent;
-  int2     window_pos;
+  float2   window_pos;
   uint32_t cursor_index;
 };
